backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
aiosqlite
greenlet
pydantic
pydantic-settings
pyjwt
supabase
langchain
langchain-openai
python-dotenv
redis
stripe
